global sidebar created by placing sidebar in components folder
