+++ src/app/(dashboard)/layout.tsx
@@ -1,0 +1,15 @@
+import { ReactNode } from "react";
+import Sidebar from "../../../components/Sidebar";
+
+interface DashboardLayoutProps {
+  children: ReactNode;
+}
+
+export default function DashboardLayout({ children }: DashboardLayoutProps) {
+  return (
+    <div className="flex">
+      <Sidebar />
+      <main className="flex-1 p-6">{children}</main>
+    </div>
+  );
+}
