+++ src/app/(dashboard)/new-task/page.tsx
@@ -1,0 +1,3 @@
+export default function NewTaskPage() {
+  return <h1 className="text-3xl font-bold">Create a New Task</h1>;
+}
