+++ src/app/(dashboard)/profile/page.tsx
@@ -1,0 +1,3 @@
+export default function ProfilePage() {
+  return <h1 className="text-3xl font-bold">Profile</h1>;
+}
