+++ src/app/(dashboard)/settings/page.tsx
@@ -1,0 +1,3 @@
+export default function SettingsPage() {
+  return <h1 className="text-3xl font-bold">Settings</h1>;
+}
