--- src/app/new-task/page.tsx
@@ -1,4 +1,0 @@
-export default function NewTask()
-{
-    return <h1>NewTask</h1>
-}
