--- src/app/profile/page.tsx
@@ -1,4 +1,0 @@
-export default function ProfilePage()
-{
-    return <h1>Profile Page</h1>
-}
