--- src/app/settings/page.tsx
@@ -1,4 +1,0 @@
-export default function SettingPage()
-{
-    return <h1>Settings Page</h1>
-}
